+++ Engine/Source/Runtime/ImageWrapper/Private/Formats/JpegImageWrapper.h
@@ -1,4 +1,4 @@
-// Copyright 1998-2019 Epic Games, Inc. All Rights Reserved.
+// Copyright Epic Games, Inc. All Rights Reserved.
 
 #pragma once
 
@@ -25,8 +25,8 @@ public:
 
 	//~ FImageWrapperBase interface
 
-	virtual bool SetCompressed(const void* InCompressedData, int32 InCompressedSize) override;
-	virtual bool SetRaw(const void* InRawData, int32 InRawSize, const int32 InWidth, const int32 InHeight, const ERGBFormat InFormat, const int32 InBitDepth) override;
+	virtual bool SetCompressed(const void* InCompressedData, int64 InCompressedSize) override;
+	virtual bool SetRaw(const void* InRawData, int64 InRawSize, const int32 InWidth, const int32 InHeight, const ERGBFormat InFormat, const int32 InBitDepth) override;
 	virtual void Uncompress(const ERGBFormat InFormat, int32 InBitDepth) override;
 	virtual void Compress(int32 Quality) override;
